fix: cypress as a dev dependency
Moves `cypress` from deps to devDeps, allowing users to set the version of Cypress they want to use in their package.json (without having to rely on a Yarn resolution to force the version).

Currently, as `cypress` is a dependency, even setting the version to `7.1.0` for example will still use the latest 6.x version installed by the schematic. With this fix, it should use the version specified in the package.json